Refactor the code to extract the filtering to a different object. The intention of the extraction is to share code between the client and server side filtering.
